Added some comments on games and cost functions for drivers based on their strategies
This contains comments regarding cost functions for the BRD file, and logic around how to come up with strategies for drivers to lower their cost and improve the overall solution.
This file will be based on social objective functions that calculate a cost for each strategy profile and return a cost, the lower the cost the higher the utility. Each driver wishes to minimise their cost and the number of high priority constraints that are broken in the solution set provided by the GA

In this case, it will return a lower cost if drivers can reduce high constraint penalties and swap with other drivers even if the payoff isn't as good for the other driver. This will hopefully reduce high priority constraints that are violated and fix broken solutions