Made changes to the structure of the BRD Class
Separated player logic over to the Player.java file, where this hold important information needed for players to participate in a BRD game.

Drivers with broken constraints, met constraints, and no constraints are added to separate ArrayLists in the Individual file and set in the Player file, so when the Improve Solution method is called, appropriate matches can be made based on a players given Strategy and their calculated utility

Further progress needs to be made on the game itself, but this is getting closer to completion. Bugs in the code I'd written prevented progress to be quicker, but these have been overcome.

TODO: Finish improve solution method, and finish constructing the BRD game. Finish constructing the run method to run the BRD 250,000 times to compare with the GA.