Added a menu in the Application file that allows you to run the GA or the BRD
This runs the GA and BRD file.
As it stands, the BRD needs more work but this is now set up in an appropriate way to do just that.

TODO: Allocate slot method : check strategy and payoff of players and allocate the slot to player 1 who is best responding to player 2's strategy. (in this case, players score higher utility if they swap high with free. The second highest utility is high with low, and the lowest utility score as medium with free.

The thinking is that this could lead to Pure Strategy Nash Equilibriums for players as they are best responding given the strategies and scores associated with other players.