Added strategy profile for players/drivers, this will help to repair the GA output
Player.java - Strategy assigns randomly based on their priority, and if they have no priority, their option is to swap with medium or high constraint violations. This sets up a way of calculating payoff, which will be added next
Still need to make matches based on week preference and strategy.

Added similar functions to initialize player variables for running the BRD algorithm independently. This is still in progress