Fixed a bug with the fitness
During simulations where the constraint number was higher, this would not reset during each run and therefore add onto itself as a result of the fixes made to the constraint check method, this is now fixed.

The fitness score now represents the solution quality and accounts for any Basil constraint violations.